% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{eu_to_uk_men}
\alias{eu_to_uk_men}
\title{Convert European shoe size to UK shoe size for men.}
\usage{
eu_to_uk_men(eu_size)
}
\arguments{
\item{eu_size}{European shoe size (numeric).}
}
\value{
UK shoe size for men (numeric).
}
\description{
Takes a European shoe size and returns the corresponding UK shoe size for men.
}
\examples{
eu_to_uk_men(42)

}
