% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{eu_to_us_men}
\alias{eu_to_us_men}
\title{Convert European shoe size to US shoe size for men.}
\usage{
eu_to_us_men(eu_size)
}
\arguments{
\item{eu_size}{European shoe size (numeric).}
}
\value{
US shoe size for men (numeric).
}
\description{
Takes a European shoe size and returns the corresponding US shoe size for men.
}
\examples{
eu_to_us_men(42)

}
