% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{eu_to_us_women}
\alias{eu_to_us_women}
\title{Convert European shoe size to US shoe size for women.}
\usage{
eu_to_us_women(eu_size)
}
\arguments{
\item{eu_size}{European shoe size (numeric).}
}
\value{
US shoe size for women (numeric).
}
\description{
Takes a European shoe size and returns the corresponding US shoe size for women.
}
\examples{
eu_to_us_women(38)

}
