% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoesize.R
\name{shoesize}
\alias{shoesize}
\title{Shoe size conversion}
\usage{
shoesize(input, system_from, system_to)
}
\arguments{
\item{input}{The shoe size in the input system.}

\item{system_from}{The size system of the input size}

\item{sytem_to}{The size system of the output size}
}
\value{
The shoe size in the output system.
}
\description{
Convert shoe sizes between different systems
}
\examples{
shoesize(40, "EU", "UK")
}
