% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{uk_to_eu_men}
\alias{uk_to_eu_men}
\title{Convert UK shoe size to European shoe size for men.}
\usage{
uk_to_eu_men(uk_size)
}
\arguments{
\item{uk_size}{UK shoe size for men (numeric).}
}
\value{
European shoe size (numeric).
}
\description{
Takes a UK shoe size and returns the corresponding European shoe size for men.
}
\examples{
uk_to_eu_men(8)

}
