% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{uk_to_eu_women}
\alias{uk_to_eu_women}
\title{Convert UK shoe size to European shoe size for women.}
\usage{
uk_to_eu_women(uk_size)
}
\arguments{
\item{uk_size}{UK shoe size for women (numeric).}
}
\value{
European shoe size (numeric).
}
\description{
Takes a UK shoe size and returns the corresponding European shoe size for women.
}
\examples{
uk_to_eu_women(6)

}
