% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{uk_to_us_men}
\alias{uk_to_us_men}
\title{Convert UK shoe size to US shoe size for men.}
\usage{
uk_to_us_men(uk_size)
}
\arguments{
\item{uk_size}{UK shoe size for men (numeric).}
}
\value{
US shoe size for men (numeric).
}
\description{
Takes a UK shoe size and returns the corresponding US shoe size for men.
}
\examples{
uk_to_us_men(8)

}
