% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{us_to_eu_women}
\alias{us_to_eu_women}
\title{Convert US shoe size to European shoe size for women.}
\usage{
us_to_eu_women(us_size)
}
\arguments{
\item{us_size}{US shoe size for women (numeric).}
}
\value{
European shoe size (numeric).
}
\description{
Takes a US shoe size and returns the corresponding European shoe size for women.
}
\examples{
us_to_eu_women(8.5)

}
