% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{us_to_uk_men}
\alias{us_to_uk_men}
\title{Convert US shoe size to UK shoe size for men.}
\usage{
us_to_uk_men(us_size)
}
\arguments{
\item{us_size}{US shoe size for men (numeric).}
}
\value{
UK shoe size for men (numeric).
}
\description{
Takes a US shoe size and returns the corresponding UK shoe size for men.
}
\examples{
us_to_uk_men(10)

}
